requirements/source.txt
# The requirements checked out as git submodules in vendor/src
# Must be kept up-to-date at all times.

# NO EDITABLE/VERSION CONTROL REQUIREMENTS ALLOWED!

python-dateutil==2.4.2

sqlparse==0.1.15

cssselect==0.9.1

Werkzeug==0.10.4

django-devserver==0.8.0

amqp==1.4.6

basket-client==0.3.11

billiard==3.3.0.20

bleach==1.4.1

bitly_api==0.3

celery==3.1.18

elasticsearch-dsl==0.0.4

elasticsearch==1.6.0

hash_ring==1.3.1

html5lib==0.99999

jingo==0.7.1

# FIXME: not released yet as of 2015-05-08
jingo-minify==0.6.0

kombu==3.0.26

oauthlib==0.7.2

python-magic==0.4.6

python-memcached==1.54

python-openid==2.2.5

pytidylib==0.2.4

# FIXME: not the newest version given errormill constraints
raven==3.3.12

requests==2.7.0

requests-oauthlib==0.5.0

tower==0.4.1

urllib3==1.10.4

URLObject==2.4.0

pyquery==1.2.9

Django==1.7.9

commonware==0.4.3

django-allauth==0.21.0

django-appconf==1.0.1

django-celery==3.1.16

django-constance==1.0.1

django-extensions==1.5.5

django-honeypot==0.5.0

django-memcached-hashring==0.1.2

django-nose==1.4.1

djangorestframework==2.4.6

django-smuggler==0.6.0

django-soapbox==1.1

django-statici18n==1.1.3

django-taggit==0.15.0

django-tidings==1.1

django-timezones==0.2

django-waffle==0.10.1

django-cacheback==0.9.1

django-debug-toolbar==1.3.2

django-mozilla-product-details==0.7.1

django-ratelimit==0.5.0

django-flat-theme==0.9.5

django-sundial==1.0.1
